requirements.txt
requests
python-dotenv
feedparser
beautifulsoup4
trafilatura
googlenewsdecoder
lxml_html_clean
